Fix libunwind-generic.so symbolic link never being created in make install
Commit 7d43108f9c41a2ccbe0adc8713523ad60469e026 ("No
libunwind-generic.so if configured with --disable-shared") introduced a
check to avoid creating a broken libunwind-generic.so link, but the
result of the commit is that libunwind-generic.so is never created (at
least when installing to a clean directory).

We need to check for the installed libunwind-$(arch).so file,
libunwind-generic.so will be the symbolic link name.