arm: properly handle functions with the noreturn attribute
This patch adjust the instruction pointer containing the return value
such that it is within the bounds of the function it refers to. This is
to ensure that we can successfully unwind through functions with
noreturn attribute set. Such functions might end up with a 'branch with
link' as the last instruction in the function and as a result, produce a
return value that falls outside the bounds of the function it belongs
to. This should enables us to get backtraces when software asserts on an
ARM platform.

Signed-off-by: Sturle Mastberg <smastber@cisco.com>
Signed-off-by: Hans-Christian Noren Egtvedt <hegtvedt@cisco.com>