There's no space for the eh_frame in synth_eh_frame_hdr.
Ben Avison (bavison@riscosopen.org) has observed that when a synthetic eh_frame_hdr is generated,
there is no space in it for the eh_frame, so the eh_frame value is written to, and later read from,
memory that is not assigned to this purpose, with unpredictable results.

This change adds a unw_word_t field to the dwarf_eh_frame_hdr type to make room for that value.