dwarf: Increase size of state to avoid buffer overrun
DW_CFA_remember_state used memcpy to overwrite state with the value
of rs_current. Unfortunately rs_current was slightly larger than state
potentially resulting in a buffer overrun. On my system this corrupted
rs_stack on a different thread.

Fix this by making the type of state match the type of rs_current. I've
also changed the memcpy call into an assignment which should ensure that
the sizes of state and rs_current match in future.