x86_64: Add fixup code if previous RIP was invalid (#120)
* x86_64: Add fixup code if previous RIP was invalid

Previous RIP could become invalid if a bad function pointer
was followed. Calling unwind from a signal frame in this case
skips the frame that called the bad pointer.

Check if the value at RSP looks to be valid in this case. It
should be the value of RIP pushed to the stack by the CALL instruction.
If it looks valid use this value as the next frames RIP else fallback
to the previous method.