Remove aux directory
aux is a reserved filename on Windows. Presence of the directory
in the repo will prevent a git checkout on Windows.

The aux directory contained GPL code related to autotools,
it appears to not be directly referenced or necessary.

Assume it was a commit error. Remove it.

Add aux to the list of autotool releated ignores in
.gitignore