mark unw_getcontext functions as returns_twice
This instructs the compiler to avoid optimizations in the callee that
would potentially interact poorly with interpreting the state, such as
frame-pointer-elimination on certain platforms, alloca lifetime
analysis/slot reuse, and so on.

It also passes along the return value from getcontext, when used (which
fortuitously is already UNW_ESUCCESS or -UNW_EUNSPEC), instead of
forcing it to be 0.