Add port for Linux on LoongArch
This commit adds support for Linux on LoongArch. Only 64-bit is supported
at the moment.

The changes were tested on a LoongArch architecture Loongson-3A5000
processor (LoongArch is a new RISC ISA, which is a bit like MIPS or
RISC-V.).

After applying the patch, the test results is as follows:
$ make check
 ...
 FAIL: Ltest-cxx-exceptions
 PASS: run-check-namespace
 PASS: run-ptrace-mapper
 PASS: run-ptrace-misc
 PASS: run-coredump-unwind
 PASS: run-coredump-unwind-mdi
 ============================================================================
 Testsuite summary for libunwind 1.6-rc1
 ============================================================================
 # TOTAL: 36
 # PASS:  35
 # SKIP:  0
 # XFAIL: 0
 # FAIL:  1
 ...

When the commit "f1cee65e7594 (Link Ltest-cxx-exceptions with libunwind)" is
reverted, all test cases can be passed.

Signed-off-by: Tiezhu Yang <yangtiezhu@loongson.cn>
Signed-off-by: Qing Zhang <zhangqing@loongson.cn>
Signed-off-by: Jinyang He <hejinyang@loongson.cn>
Signed-off-by: Youling Tang <tangyouling@loongson.cn>