arm64: Add support for VG register in signal frames
Allow local and remote unwinding through signal frames
when variable length SVE registers are pushed onto the
stack.

The vector length is saved by the kernel into the signal
context, but it is not scaled in the same way that
dwarf expects it to be. Therefore a conditional has
been added to tdep_access_regs() that scales the value
depending on whether it is going to be accessed through
a register or through memory.

Signed-off-by: James Clark <james.clark@arm.com>
Change-Id: Ie16aa22b36127ba5aa81f20280b1df7e4ba6d49b