arm64: Add test for SVE unwinding from signal handler
The test is skipped if the compiler doesn't support
SVE, otherwise it uses some SVE intrinsics across a
few functions and then sends a signal to take a
backtrace from the sum() function.

Without the previous SVE changes, the unwinding
fails at sum() and the stack looks like:

-> signal_handler()
->   kill()
->     sum()

With the changes the full stack to main can be
unwound and the test passes:

-> signal_handler()
->   kill()
->     sum()
->       square()
->          main()

Currently this test will pass on a Graviton 3 instance
on AWS and with gcc-10.

Co-authored-by: Kent Cheung <Kent.Cheung@arm.com>
Signed-off-by: James Clark <james.clark@arm.com>
Change-Id: Ibe38e6b0fa26276c545f044ffdd26fbdb2789c38