Correct build failure
The following pull request, targeting aarch64, introduced a build
failure:

  https://github.com/libunwind/libunwind/pull/360
  (f67ef2867bc5e74cdadb3acb790c3fc6161b22e9)

The failing build was with the msvc compiler.  According to

  https://docs.microsoft.com/en-us/cpp/assembler/inline/inline-assembler?view=msvc-170

the compiler only supports inline asm on x86.

Since the inline assembly is only for local unwinding, this patch fixes
the build by only compiling the inline assembly if UNW_REMOTE_ONLY is
not defined.  This works because the msvc build has UNW_REMOTE_ONLY set
in the CMakeLists.txt file.