Find segment containing .text by flag not order
The existing code was identifying the segment containing the executable code
(.text section) by assuming it was effectively the first loadable segment. Th
has always worked up until recently but was an invalid assumption. The right
thing to do is check the segment flags to see if the execute bit is set.

This change does just that. Tested on Linux x86_64, no new regressions.