[LoongArch64] Reflect the kernel ABI changes since the port went upstream
The original port was done with an early in-house port of Linux that, in
addition to slightly different UAPI headers, also featured a MIPS-like
ABI (the so-called "old world" ABI). The upstream ABI has been revised
since long ago and already frozen, so adjust the port for the "new world".
In particular, we don't have the 24-byte signal trampoline area any more
which was a MIPS o32 thing.

We don't need to keep compatibility for the old-world kernels, because
distributions using said kernels invariably packaged their own libunwind
fork with corresponding support, and the few users tracking upstream
kernels should all have moved on.

Fixes: c5f1d12c77de ("Add port for Linux on LoongArch")
Signed-off-by: Youling Tang <tangyouling@loongson.cn>
Signed-off-by: WANG Xuerui <xen0n@gentoo.org>