get_elf_image: Always prefer passed path buffer
If the caller passes a path buffer, always use it. If the buffer is too
small, fail. As the returned value is of no use to the caller anyway.
Only allocate a buffer, if NULL was passed.

Makes this function guaranteed AS-safe as long as a path buffer is
provided. Which also improves AS-safety of the callers, which includes
`unw_get_proc_name`.